requirements.txt
click==8.1.7
numpy==1.22.4
pandas==2.2.1
Pillow==10.2.0
scikit_learn==1.4.1.post1
scipy==1.8.1
torch==1.12.0+cu113
torchvision==0.13.0+cu113
tqdm==4.66.2
